Fix over-engineering: remove dead code, extract components, simplify deps
- Remove dead code: _check_coherence() (Hierarchy Manager), severity_badge()
  (Streamlit utils), AuditEventResponse (API schemas)
- Add useAsyncOperation hook to replace 8+ repeated try/catch/finally patterns
- Extract 11 domain components from monolithic page files (forecast, sop,
  data-onboarding, sku-transitions) reducing page sizes by 60-85%
- Refactor S&OP Meeting export buttons from 3 duplicate handlers into
  config-driven loop
- Remove get_app_state wrapper from API routers — use Request directly
- Update CLAUDE.md, README.md, CONCEPTS.md with new component structure
