+++ CLAUDE.md
@@ -57,7 +57,7 @@ forecasting-product/
 │   ├── api/                # FastAPI REST endpoints (auth-protected)
 │   │   ├── app.py          # create_app() factory, core endpoints, router registration
 │   │   ├── schemas.py      # Pydantic request/response models
-│   │   ├── deps.py         # Shared dependencies (get_app_state)
+│   │   ├── deps.py         # Shared dependencies
 │   │   └── routers/        # Domain-specific API routers
 │   │       ├── series.py       # GET /series/{lob}, POST /series/breaks, cleansing-audit, regressor-screen
 │   │       ├── hierarchy.py    # POST /hierarchy/build, aggregate, reconcile
@@ -123,7 +123,11 @@ forecasting-product/
 ├── frontend/               # Next.js 15 frontend (TypeScript, Tailwind, Recharts)
 │   ├── src/app/            # App Router pages (login + 8 workflow pages)
 │   ├── src/components/     # Reusable components (charts, AI panels, layout, shared)
-│   ├── src/hooks/          # React Query hooks for each API endpoint
+│   │   ├── forecast/       # Decomposition, comparison, constrained forecast panels
+│   │   ├── governance/     # Model cards, lineage, BI export panels
+│   │   ├── pipeline/       # Multi-file analysis, pipeline execution panels
+│   │   └── sku/            # SKU mapping, override management panels
+│   ├── src/hooks/          # React Query hooks + useAsyncOperation utility
 │   └── src/lib/            # API client, auth, types, constants
 ├── tests/                  # 1060+ tests (pytest)
 ├── configs/                # YAML configuration files
